projeto-temp/requirements.txt
cryptography
eventlet
flask
flask-socketio
flask-sqlalchemy
matplotlib
openpyxl
pandas
psutil
pyopenssl
python-dotenv
pytz
reportlab
requests
schedule
sqlalchemy
werkzeug
xlsxwriter
cryptography
eventlet
flask
flask-socketio
flask-sqlalchemy
matplotlib
openpyxl
pandas
psutil
pyopenssl
python-dotenv
pytz
reportlab
requests
schedule
sqlalchemy
werkzeug
xlsxwriter
cryptography
eventlet
flask
flask-socketio
flask-sqlalchemy
matplotlib
openpyxl
pandas
psutil
pyopenssl
python-dotenv
pytz
reportlab
requests
schedule
sqlalchemy
werkzeug
xlsxwriter
cryptography
eventlet
flask
flask-socketio
flask-sqlalchemy
matplotlib
openpyxl
pandas
psutil
pyopenssl
python-dotenv
pytz
reportlab
requests
schedule
sqlalchemy
werkzeug
xlsxwriter
